+++ crates/ewwii/src/widgets/build_widget.rs
@@ -51,7 +51,7 @@ fn build_gtk_widget_from_node(
             build_tooltip(props, children, widget_reg)?.upcast()
         }
         WidgetNode::LocalBind { props, children } => {
-            build_localbind(props, children, widget_reg)?.upcast()
+            build_localbind_util(props, children, widget_reg)?.upcast()
         }
         WidgetNode::CircularProgress { props } => {
             build_circular_progress_bar(props, widget_reg)?.upcast()